+++ popup/src/app/providers/NotificationProvider/model/types/types.ts
@@ -1,0 +1,3 @@
+import { AlertProps } from "@mui/material/Alert/Alert";
+
+export type Severity = AlertProps["severity"] | null;