+++ popup/src/app/config/i18n.ts
@@ -40,7 +40,7 @@ i18n
     fallbackLng: "en",
     supportedLngs: supportedLanguages,
     nonExplicitSupportedLngs: false,
-    ns: ["messages", "modals", "settings", "customize", "dictionary", "common"],
+    ns: ["messages", "modals", "settings", "customize", "dictionary", "about", "common"],
     backend: {
       loadPath: `${baseUrl}_locales/{{lng}}/{{ns}}.json`,
     },