+++ popup/src/app/config/routerPath.ts
@@ -2,6 +2,7 @@ export enum AppRoutes {
   SETTINGS = "settings",
   CUSTOMIZE = "customize",
   DICTIONARY = "dictionary",
+  ABOUT = "about",
   LOGIN = "login",
 }
 
@@ -9,5 +10,6 @@ export const RouterPath: Record<AppRoutes, string> = {
   [AppRoutes.SETTINGS]: "/",
   [AppRoutes.CUSTOMIZE]: "/customize",
   [AppRoutes.DICTIONARY]: "/dictionary",
+  [AppRoutes.ABOUT]: "/about",
   [AppRoutes.LOGIN]: "/login",
 };